am62ax: ipc: Suspend C7X based on the next host state
Use the next host state to decide what to do during suspend, whether
C7X should be shutdown or not.
This support is broken currently as the general suspend task is secured
with MCU ONLY LPM macro that is defined only for MCU R5s. This results
in C7X core always posting the task that was never created and hence the
timeout on remoteproc side waiting for ACK.

Fixes: SITSW-5779

Signed-off-by: Akashdeep Kaur <a-kaur@ti.com>