allsocs: sci: Fix true and false to boolean
Fix for the following two MISRA C issues:
 - MISRA.ETYPE.CATEGORY.DIFFERENT.2012
 - MISRA.STMT.COND.NOT_BOOLEAN.2012

Only in ti-cgt-c7000_4.1.0.LTS compiler, "true"
and "false" are not boolean. So, this causes
a type mismatch when a boolean is expected
in if statments and while loops

To fix this, all areas where "true" and "false"
are used in if conditions and while loops have
been casted as bool. This way it gaurantees
that "true" and "false" will be of type bool,
no matter the compiler

Fixes: SYSFW-7738

Signed-off-by: Dillon Frank <d-frank@ti.com>