allsocs: sci: Fix memory boundary access
Fix ABV.ANY_SIZE_ARRAY and ABV.GENERAL

"Buffer Overflow - Array Index Out of Bounds"

The data type of "pReqPrm->flags" is uint32_t,
but "header->flags" is of type uint8_t. This
cast would usually be fine, but since we are
using the size of "pReqPrm->flags" as the
boundary of the for loop, and then using
that index to iterate over "header->flags"
there is a chance of the index being larger
than what "header->flags" can support. So,
the for loop boundary just needs to be
modified to account for the type difference
by dividing by uint8_t.

Fixes: SYSFW-7738

Signed-off-by: Dillon Frank <d-frank@ti.com>