all: lpm: Add suspend and resume hook
During low power mode entry the DM R5 is suspended and resumed back
during the LPM exit. Along with the DM, the DM application also
needs to suspend and resume. Add hooks in the DM suspend and resume
handler so that the application running along with DM can register
suspend and resume handlers.

Fixes: SITSW-5863

Signed-off-by: Sebin Francis <sebin.francis@ti.com>