allsocs: device_manager: Fix flag check for response
All TISCI messages intended for TIFS which have
SCISERVER_FORWARD_MSG flag have TISCI_MSG_FLAG_AOP
along with the existing flags in the header.
This change was made in commit 47f33021c58 for
SITSW-4773.

But the flag that is being checked after sending
the message to TIFS is the request parameters flag
not the header flag which has the AOP flag.

Remove the SCISERVER_FORWARD_MSG check in polling
mode and fix the flag in both polling and interrupt
mode.

Fixes: SITSW-3685

Signed-off-by: Sparsh Kumar <sparsh-kumar@ti.com>