am62/a/p/d: dm: Enable interrupt mode for DM2TIFS forwarding queue
Add an API to update the sciclient operating mode to
interrupt mode. Initialize the required semaphore and
ISR handler for the interrupt mode. During boot initialize
the sciclient in polling mode. Then update the sciclient
to interrupt mode in the sciserver Initialization. This is
done to make sure that the scheduler has started before
running sciclient in interrupt mode, because of the
scheduler dependency on the semaphore.

Guard the API call with "CONFIG_DM2TIFS_INTERRUPT_MODE"
which is disabled by default. It can be enabled in the
project file of the soc if the user wants sciclient in
interrupt mode.

Fixes: SITSW-2625

Signed-off-by: Sparsh Kumar <sparsh-kumar@ti.com>